Render idef0 diagram using webapp
The webapp receives the URL pattern /svg/<string> and renders the
idef-0 diagram in svg format.

The string is a base64-encoded and deflate compressed in accordance to
the PlantUML specification.

Once the base64 string is decoded by the webapp, the custom DSL
(vorachet/FindFuncDSL) is then parsed and translated into
Jimmyjazz/IDEF0-SVG DSL syntax. The SVG image illustrating the IDEF0
diagram is then rendered and returned as the web response.